jetson/requirements.txt
opencv_python_headless==4.2.0.34
imutils==0.5.4
facenet_pytorch==2.5.2
google-cloud-storage==1.43.0
python-firebase==3.0.1
Jetson.GPIO==2.0.17
Pillow==8.4.0
torch==1.10.1
torchvision==0.11.2
